web/requirements.txt
Django>=1.11,<1.11.99
psycopg2-binary
pytz

requests
djangorestframework
django-filter
django_extensions
